chore: rename hook files
