+++ frontend/src/modules/hooks/userIsMobileDevice.ts
